[FIX] documents: behavior of mail-chatter-dropzone
Steps to reproduce:
1. select a document
2. open chatter
3. drag(preview part) and drop to the others folder
4. the mail-chatter-dropzone remains open

Technical Reason:
Removing the unnecessary stop propagation on the search
panel's document drop event as it was preventing it from
being caught by the window keeping the chatter drop zone opened.

After this commit:
mail-chatter-dropzone will be closed after moved to drag documents
to the folders.

Task-4268738

closes odoo/enterprise#73368

X-original-commit: d6c10e2e9557e8f219403396d188424c0973ff6d
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>
Signed-off-by: Astik Singh (assi) <assi@odoo.com>