[IMP] account_reports: batch label builder
When using a custom groupby, we can use a label builder using the grouping key.

Before it was called once for each grouping key and some optimization might have not been done as we could not batch it.

The new way call the label builder once with all the grouping keys. This way it is batched and we can make optimization if required.
The label builder function returns a dict with each grouping key containing the name of the line.
The order of this dict is important as it will be the order rendered on the report.
So another feature allowed by this is that we can manually order grouping keys and therefor lines.

Part-of: odoo/enterprise#72675
Related: odoo/upgrade#7798
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>