[FIX] project_enterprise: avoid creating employee in test
Before this commit, `:TestSmartSchedule.test_multi_users_tasks` test in
`project_enterprise` module creates 2 employees but the problem is `hr`
module is not in the dependencies of `project_enterprise` and so that
test fails when `hr` is not installed.

This commit adapts the test to make sure to not create any employees
inside that test.

runbot-76567

X-original-commit: eb35533c240d3b6ed2b5b9331ef56f412e15b01f
Part-of: odoo/enterprise#72945
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>