[FIX] project_enterprise_hr: adds test_smart_schedule test
Before this commit, the `test_multi_users_tasks` test defined in
`/project_enterprise:TestSmartSchedule` failed when `hr` module is not
installed since that test create 2 employees.

This commit moves the test in `project_enterprise_hr` to make sure the
test works as expected and we can create employee since `hr` is in the
dependencies of `project_enterprise_hr` module.

runbot-76567

closes odoo/enterprise#72945

X-original-commit: fbbb2d439137c9169d373ed7252d4571148a66a4
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>