[FIX] planning: define full time required hours for the employee in test
Before this commit, since the `full_time_required_hours` field is unset
on the calendar of Aaron employee, the test `/planning:TestUi.test_01_ui`
fails because no progressbar is displayed for that employee in the gantt
view of `planning.slot` model, meaning the employee is not supposed to
work.

This commit sets `full_time_required_hours` on the calendar of `Aaron`
employee to display the progressbar for that employee as it was expected
before the rework of flexible hours for resources.

runbot-99111

closes odoo/enterprise#72807

X-original-commit: f6223e028faaacadbc06494bb0d087753ab36e1a
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>