[FIX] l10n_br_avatax: send correct hsCode for services
Some cities that process service invoices [1] have stricter hsCode
validations than others and error on the hsCode we send.

Avalara requires periods to be removed from the hsCode field for goods
transactions. This was already the case. For services they inform us
that the periods should be kept but leading zeroes should be stripped
[2].

[1] e.g. Belo Horizonte
[2] There's no documentation on this. This was specified by Avalara
support in an internal support ticket.

opw-4009684

closes odoo/enterprise#72857

X-original-commit: c48171973d44778f9846604068d1982005f322a3
Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>
Signed-off-by: Joren Van Onder (jov) <jov@odoo.com>