[IMP] sign: avoid reopening PDF form after sending for signature
Before this commit, when a PDF was sent to users for signature, the form would
reopen automatically after sending, which could disrupt the user flow.

With this commit, the form no longer reopens after sending the document.
However, users can still access and view the sent document by clicking on the
link or the corresponding kanban card.

task-4184028

closes odoo/enterprise#72474

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>