[FIX] sign: test_ui
The shared_sign_request_tour, attempts to open some template and then click on some sign item.
However it fails when it tries to click on the sign item, because it's not found.

This happened because the tour sometimes open the wrong sign template (one with no items).
This commit aims to fix the issue by archiving other sign templates when testing this flow.

fixes runbot-98036

X-original-commit: e1067759a4d4aafefd8bb5cfc698596f75d33509
Part-of: odoo/enterprise#72840
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Ahmad Almaghraby (alah) <alah@odoo.com>