[FIX] mail: add a reaction icon in public page
Before this commit:

On mobile, when a guest joined a public channel via an invitation URL, the
'Add a Reaction' icon appeared as [] instead of the oi-smile-add icon. This was
due to Font Awesome loading after odoo_ui_icons, causing a conflict with the
fa class.

After this commit:

This commit fixes the issue, and now the 'Add a Reaction' icon displays correctly.

closes odoo/enterprise#72659

X-original-commit: cda19f229a510a60be53bd58bba369b35b4c60bc
Related: odoo/odoo#184996
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>