[IMP] account_reports: Added a caret builder function on the custom groupby map
This enables us to set custom caret_options on specific grouping keys when utilizing a custom groupby

For instance we have a custom groupby without model, and we need the caret feature.
Using the example of adding a custom line of initial balance using a string grouping key, we might still need caret options for the other lines from the current groupby.

closes odoo/enterprise#72216

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>