[IMP] account_reports: Allow using string grouping keys in a custom engine.
To use this feature we need a custom groupby using a custom groupby map with a label builder or the model set to None.
With this we can add specific lines using string grouping keys then process them in a postprocessor.

For instance adding a balance line with the grouping key `balance_line` and specific values related to the balance in the columns.
Then in the postprocessor, we can use this to set up an initial line containing the balance.

Part-of: odoo/enterprise#72216
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>