[IMP] {hr_payroll_}account_iso20022: define ChrgBr on payment level
Currently for batch payments using an ISO 20022 payment file, we allowed
defining a charge bearer (ChrgBr) at the batch level. However, it
happens that we want to send payments in a batch where we want some of
the payments to have a different charge bearer than others.

In order to allow this, we move the charge bearer from payment batches
to individual payments. If a batch contains payments with different
charge bearers, we define the `ChrgBr` node on the individual
`CdtTrfTxInf` nodes in the XML file. If all payments use the same charge
bearer, we keep the old behavior and leave the `ChrgBr` at the `PmtInf`
level.

Bank journals now also get the option to define a default charge bearer
for all payments created from that journal.

task-4275038

closes odoo/enterprise#72972

Related: odoo/upgrade#6688
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>