[IMP] Recruitment: Request Signature
Problem
----------
- In an application form view, There are 2 Signature Request wizard accessible from the cogwheel menu. We have to use only the wizard from sign. And not our custom one in hr_recruitment_sign.
- Change this sign wizard to replace the user mail with the applicant mail.
- In case of selection of multiple application in list view. We have to be able to send signature request to all of them with a new wizard in the action cogwheel menu. This wizard must be the same as the sign request for employee one.

Objective
----------
- Remove our sign request wizard in the cogwheel menu of applicant form view.
- Adapt the sign wizard to display the applicant email by default.
- Add the multiple sign request wizard in the cogwheel menu of list view.

Solution
----------
For the multiple sign request : I use the previous customized wizard in hr_recruitment_sign (replaced by the sign one). Just adapt it to be usable with multiple records.
So now the "signature request" action available in the cogmenu is now display only in list view for multiple sign requests.
Finally, for the sign request wizard in form view. To make it custom for the hr_applicant model. I needed to adapt and override the context of the cogmenu item to add the default signer and the active model. And adapt and override the sign_send_request to change the default signer.

See odoo/upgrade#6694
task-4260572

closes odoo/enterprise#73040

Related: odoo/odoo#194481
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>