[IMP] marketing_automation_website_sale: add anniversary campaign
Add a "1 year anniversary" campaign that also serves to demo
dynamic domains inside campaigns

The anniversary campaign sends a mailing to all contacts who
did not yet exist when the campaign was created, but are now
"1-year-old" in the database at the time the domain is reevaluated
to add participants. Users are expected to update the mailing with
relevant information such as a valid coupon and otherwise personalized text.

It is important that the context be `older than one year` and not strictly
`exactly one-year-old` as we want the domain to always be valid in the future
in case the next action should occur "in X days". As the domain will be
re-evaluated at that time, and may otherwise not match the partner anymore.

This is not intuitive but it is a natural consequence of automatically
merging domains with '&' which can't be changed easily. We thus prefer
working around the issue, as is done here.

task-4240689

Part-of: odoo/enterprise#72570
Related: odoo/odoo#184830
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>