[FIX] account_accountant: Add period condition in Deferred Expense Report
Version: 17.0+

Issue:
In the Deferred Expense Report, there is a missing check for if the `period_end` is before the `period_start`
for when the deferred method is 'day'.
If the `period_end` is in the same month as selected month of the report,
the later column will compute the amount to be negative of the `amount_per_day` times the difference of days.
For example, if the end date is 2024-11-04 and we look at the month of November,
the later column will display the negative amount of 26 days.

Purpose of this PR:
To add an additional condition if the `period_end` is before the `period_start`then the amount should be 0.

opw-4231549

closes odoo/enterprise#73105

X-original-commit: 285fa92c75c8fecc7b1dd4d3b9e25b77e095cfde
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Ian Dao (iada) <iada@odoo.com>