[IMP] pos_restaurant_preparation_display: display floating order name on card
Before this commit:
====================
Only the tracking number was shown on the kitchen card for floating orders. If
the user changed the name of a floating order, it could create confusion if the
displayed name was different.

After this commit:
==================
The floating order name will be displayed for floating orders, otherwise, the
tracking number will be shown.

task- 4274465

closes odoo/enterprise#73951

X-original-commit: 0319abf46f8ad64593de0545da096ea5c9c390a2
Related: odoo/odoo#187474
Signed-off-by: David Monnom (moda) <moda@odoo.com>
Signed-off-by: Jitendra Kumar Prajapat (jipr) <jipr@odoo.com>