[IMP] appointment: prevent drag when groupby type
Prevent being able to drag an appointment type when kanban
is grouped by Type (category field) as this changes the
appointment type category and messes up the fields display.

Setting the category as readonly on the kanban view to remove
the possibility to change it on groupby via drag and drop.

Task-4282969

closes odoo/enterprise#73239

X-original-commit: 5755c56392fe31d0be4c7bdc09c7a8b055b55351
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Amélie Dieudonné (amdi) <amdi@odoo.com>