[FIX] project_sale_subscription: end reccurence on sub closing
Steps:
Create a product of type service that creates a task.
Create a subscription with that product, and confirm.
Note that if you set the task's state to done, it creates another one.
Close the subscription.

Issue:
If you set the task's state to done, it still creates another one,
whereas it would make sense that the task's recurrence ends with
the "invoice recurrence".

Cause:
Actually, it was first designed that way, and `_set_closed_state`
was overridden to unlink the recurrence.
But then that method and `set_close` were rethought and the override
wasn't called anymore.

Solution:
Just do the same thing, but in `set_close` i/o `_set_closed_state`.

task-4259561

closes odoo/enterprise#73436

X-original-commit: 57d0f514a681ab095548ec7684bdda777af4f896
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>