[FIX] social_instagram: instagram post links to account profile
Issue: After posting content through Social Marketing on Instagram,
the generated link redirects to a non-existent post or page.

Steps to Reproduce:

1. Install the Social Marketing tool.
3. Set up an Instagram account connection.
3. Create a post and attempt to click on the generated link for
redirection to the post on Instagram.

Solution:

The issue stems from using Instagram's internal post_id, which does not
work for direct URL access to the post (we will need the shortcut or the
post permalink which is not available without doing extra calls). The
fix will then focus on providing a link to the account's profile from
the post.

opw-4243329

closes odoo/enterprise#76313

X-original-commit: 7d30f86c8ea454fbcfeadff001a4dabc71d37122
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>