[FIX] timesheet_grid: check unit_amount encoded only for timesheets
Before this commit, when the user alters the unit_amount field of an
analytic account line which is not a timesheet and the amount has more
than 6 digits then the user will get a user error saying he cannot
encode a number with more than 6 digits for the unit amount field.

This commit makes sure the check made on the unit amount is only made
when the analytic line altered is a timesheet.

closes odoo/enterprise#72839

X-original-commit: 057d2eefeb4bb94222f7ddac17dc572ab65d63d4
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>