[FIX] planning: fix 'requests to switch' filter's timezone problem
The 'Requests to switch' filter is supposed to show slots that are
unwanted and that are not in the past.

Prior to this shift, some shifts that were in the future were not
shown by the filter.

That was because the planning.slot field end_datetime was encoded
in UTC while the datetime.datetime.now() function in the filter's
domain was returning the datetime in the client's timezone. Shifts
in the time difference between the user's timezone and UTC would be
mistakenly marked as "past shifts" while they might still be in the
future.

We fix this by using the 'to_utc()' function which should make the
filter domain's comparisson between 2 UTC datetimes.

task-4286161

closes odoo/enterprise#73442

X-original-commit: 00f56f3f90b869dc9fb11be58c81fbd8ec577a74
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>