[IMP] hr_payroll_expense: Add Error when no expense rule is found
This commit improves "expenses to be reimbursed in a payslip "config

Before this commit:
There was no way for a user to know if a payslip set to be reimbursed
in a payslip would never be due to a bad configuration.

After this commit:
The user now gets an error when choosing to reimburse an expense
through a payslip, asking to first setup a proper rule.

closes odoo/enterprise#74649

Task-id: 4273421
X-original-commit: b1ec60e03767d35fad09bd308e8ef2f4babb6e42
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>