[FIX] account_loans: add dependency to base_import
Steps to reproduce:
- Install account_loans
- Go to Settings
- Disable "Import & Export" option
- Save

Issue:
The db is not accessible anymore because of the following error:
-----------------------------------------------
The following modules are needed by other modules but have not been defined,
they may not be present in the correct asset bundle:
- @base_import/import_action/import_action
- @base_import/import_model

The following modules could not be loaded because they have unmet dependencies,
this is a secondary error which is likely caused by one of the above problems:
- @account_loans/components/loans/import_action
-----------------------------------------------

Cause:
When "Import & Export" option is disabled, "base_import" module and many
others are unsintalled because they depend on "base_import" module.
"account_loans" is not uninstalled but it should be because it depends on
"base_import" module as it extends its "ImportAction" component.

opw-4283944

closes odoo/enterprise#72983

X-original-commit: 5723e16d6084f9626a860883ee77461768c27436
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>