[IMP] pos_*: change table name system
*: l10n_de_pos_res_cert, pos_restaurant_appointment,
pos_restaurant_preparation_display

Before, the table system was very basic and not very convenient for the
user. When he wanted to create new table between floor plan, sometime
several tables where created with the same number. This was due to the
fact that the system was not able to detect the last table number
between all floor plan.

Now each floor plan has a prefix and the system will name new table with
the next number of the last table of the same floor plan.

closes odoo/enterprise#72878

Taskid: 4283519
Related: odoo/odoo#185449
Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>