[IMP] pos_restaurant_appointment: booking flows better ux
In this commit

We are enhancing the user experience for the table booking flow at the
restaurant. Questions added to the website during the table booking process will
now be displayed in the booking form. We will no longer show appointments with a
"No Show" status. We have also improved the Gantt view for managing bookings.

task-4283805

closes odoo/enterprise#73265

Signed-off-by: David Monnom (moda) <moda@odoo.com>