[REV] *: remove total below section
This reverts commit fca441f7fe024722a043cd9300dbaae3a2a2c1e0.

We got reports that the "Totals Below Sections" setting is actually important in the US and LATAM countries. The usage is there to always display amounts like that on reports, and there is a lot of resistance to softwares not allowing that. We hence reintroduce this feature.

closes odoo/enterprise#72930

Related: odoo/upgrade#6678
Signed-off-by: Quentin De Paoli (qdp) <qdp@odoo.com>