[FIX] helpdesk{_timesheet, _sale_timesheet} : Show all non-billable projects
Current behavior before PR:
When creating a Helpdesk team and the Timesheets option enabled and
Time Billing disabled you will not get all non-billed projects in the dropdown menu.
This is happening because the domain for project_id is checking the company_id
to be the same as the company_id of the team.
https://github.com/odoo/enterprise/blob/18.0/helpdesk_sale_timesheet/models/helpdesk_team.py#L10
Though since this commit https://github.com/odoo-dev/odoo/commit/16a07ed1bfd33a193b363f3e87aab3f9bd6f8bd0 company_id
is not required for projects. Since the domain of the project_id field was introduced
here https://github.com/odoo/enterprise/pull/4956/commits/c82b1b019cebd1f78749b9563e717fae88d7fae5 and the reason why we check the company_id
is not the same anymore since the commit I mentioned above we should not
use the same domain.

Also in this PR we are solving another issue which is where you enable
both Timesheets and Time Billing options and choose a project that has
both options enabled too then you disable the Time Billing option for the team
the project that we chose before will be kept and have both options enabled and you can
save it.

Desired behavior after PR is merged:
We are now checking for the project of the team if the company_id is False
or equal to the team's company_id.

Also when disbaling the Time Billing option we set the project_id to False.

opw-4275647
opw-4148207

closes odoo/enterprise#73433

X-original-commit: b850e8e214061526f4201c23ce3066e45d7ade43
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>