[FIX] l10n_br_avatax: Avatax BR not called from webshop
With a BR Company Setup:
Settings > Website, assign 'My Website' to BR Company
Configure Avatax Brazil in General Settings > Taxes > Avatax
Publish the Product "Regular Consumable Product" to be available in the website.
Open the web shop as Public User
Add "Regular Consumable Product" to cart
Go to checkout
Fill valid Brazilian customer info
Pay

Issue: Error will popup
```
odoo.http: Odoo could not fetch the taxes related to S00040.
It is not possible identify the CFOP
```

This occurs because the sales order is missing the "Goods Operation Type"
and the call to Avatax is not done

opw-4239483

closes odoo/enterprise#73343

X-original-commit: c675f0ffdc935171ff75fdcdab39f9be71531d70
Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>
Signed-off-by: Andrea Grazioso (agr) <agr@odoo.com>