[IMP] documents: add (copy) to duplicated folders' names
Purpose
=========
Previously, when duplicating a folder or shortcut, the duplicated records used
the same name as the original. This was confusing for the users.

This commit resolves the issue by appending "(copy)" to the name of any duplicated
parent, folder or shortcut. This ensures better clarity and avoids confusion for the users.

Task-4283698

closes odoo/enterprise#73032

Signed-off-by: Stéphane Debauche (std) <std@odoo.com>