[FIX] stock_barcode: very slow multi SN scan
Context
=======

The RFID and multiscan are able to do a lot of barcode by second.
However the quant model is old and process everything one by one.
So at each line, it does an RPC to find the related quant and display
its data.

But in the case of 1000 products in seconds, it will do 1000 RPC and
be a bottleneck for performances. Also everything is already batched
before process barcode, so we have everything ready to batch RPC.

This PR will fetch all the quants related to products in barcodes.
We add them in the cache and when we need it, we look in the cache
without RPC.
It adds an extra layer of cache for quant with the basic search keys
product, location, lot, package, owner.

Another similar issue happens when scanning a lot of Serial Numbers in
one barcode (multiple barcodes can be scanned if each individual
barcodes are separate by the separator character, which is ',' or `;` by
default.)

Steps to reproduce
==================

1. Create a storable product tracked by SN.
2. Create a delivery order of 1000 units.
3. In Inventory > Operations Types, for the Delivery Orders type,
   check "Create New" under Lot/Serial Number.
4. Open the barcode app
5. Scan a QR code containing 100 serial numbers.
> makes far too many RPC calls taking ages to resolve in 3G or slow 4G

Cause of the perf issue
=======================

You will observe that the barcode scanned is treated by the
`processBarcode`, generates a list of 100 valid barcodes and that each
of these is then processed individually here:
https://github.com/odoo/enterprise/blob/b429e4da99edc73b5cf451c05dba120dde5b3895/stock_barcode/static/src/models/barcode_model.js#L650-L674
However, each individual `this._processBarcode` will lead to two RPC
calls, one during the `_getMissingRecord` of the
`prefilled_owner_package_stock_barcode`.

To avoid this issue, we try to fetch all needed records in once but
doing that:
1. When the SN are scanned, the default barcode nomenclature parse them
   as products. Thus, the first call to
   `test_barcode_create_serials_in_batch_with_single_scan` will search
   products for those barcode but find noting;
2. Seeing we do a search for some barcode, we compare the barcodes' list
   length and the found record's length. If there is a difference (and
   it's the case here since no products match the barcodes), we list all
   not found barcodes and prepare a second search;
3. The second search is done for no particular model so barcodes are
   searched for every Barcode app's compatible models. Since we look for
   those barcodes with no model, we mark in the lazy barcode cache they
   were searched;
4. When processing individually each barcode, when checking if there is
   a serial number record for it, we saw in the cache they were already
   searched and so, if not in the lazy cache, that means they don't
   exist and no additionnal fetch is needed.

For the 3., the second search is done thanks to the parameter
`forceUnrestrictedSearch`, this parameter is used only when the used
barcode nomenclature is not the GS1. Since the GS1 nomenclature is
really precise, barcodes are correctly parsed, it's not the case with
the Odoo default nomenclature where by default, all unrecognised
barcodes are parsed as a product's barcode.

There was also another similar issue with
`prefilled_owner_package_stock_barcode` method who get called for every
SN to know if it already exist a quant for this SN (and in such case,
get back information like its package or its owner.) To avoid to call
this method for every SN, we fetch on time all the quant for the given
product (and location if it's relevant) and then, we filter those quants
by their SN in the JS.

Test Issue
==========

To manually test the flow, you can run the test
`test_barcode_create_serials_in_batch_with_single_scan` with
`debug=True` and then, in the network tab of the browser devtools, set
throtlling to 3G while the debugger of the first step of the tour is on.

Test written by: lase-odoo <lase@odoo.com>

opw-4410797

closes odoo/enterprise#77790

X-original-commit: 26f9015a885fcf192265a455e4ba9dc9ff2a39f3
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Steve Van Essche <svs@odoo.com>
Co-authored-by: svs-odoo <svs@odoo.com>