[IMP] sale_subscription: remove duplicated close reason
The close reason with id 'close_reason_3', is currently the same reason
as the 'close_reason_end_of_contract'. To avoid having subscriptions
with the same functionnal reason in two different reasons, we merge
these two.

We keep the one with the id 'close_reason_end_of_contract' as it is
referenced in the code base.

closes odoo/enterprise#73044

Task-id: 4294258
Related: odoo/upgrade#6695
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>