[FIX] pos_settle_due pos_settle_due, pos_urban_piper: order creation
`add_new_order` method is now synchronous and returns the new order.
This allows to directly use the new order in the next steps without the
need to await the promise.

The `pos_order` reference is now added to the order by calling
`getNextOrderRefs` that's isn't awaited anymore. That's means that
during a small amount of time, the order didn't have following fields:
`pos_reference`, `tracking_number` and `sequence_number`

closes odoo/enterprise#73050

Related: odoo/odoo#185822
Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>