[REF] iot: split device and box models in two files
IoT Box/IoT Device models were placed in the same file, we separated them.
Additionally, the IoT Channel model, which was placed in the same file too, has been moved to the `ir_websocket` one,
and the "IoT Connection Error" dialog have been moved to a `dialogs/` directory for more clarity in the `iot` module.

closes odoo/enterprise#73065

Signed-off-by: Quentin Lejeune (qle) <qle@odoo.com>