[IMP] web_gantt: simplify gantt pill display name
In the past, there used to be just "static" scales
(day, week, month, year) in the gantt view.
For a given scale, we displayed the start and end dates
of a record when it spanned across several units of that scale,
because then, you couldn't see both limits at once.
For example, when the scale is in weeks and the current focus
is this week, we display the dates of a record if
it starts this week and ends on next week,
because you can't see the end of it.

Now that it is possible to set the start and end date
of the period that is displayed more dynamically,
we want to keep this rationale "we display the dates
if you can't see them graphically".
So we display the start and end dates of the record
if it starts before the start of the period or
ends after the end of the period, no matter the "scale".

task-3974394

Part-of: odoo/enterprise#73086
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>