[FIX] documents: noupdate records
server actions that use document tag should be noupdate as well
because document tag records are noupdate, and they could be deleted
by clients.
possible steps to reproduce:
1) make a database 17.0
2) delete document tags
3) upgrade the database to 18.0

Document tags and server actions should be recreated in upgrade scripts
when upgrading to 18.0 if they are missing. Setting the noupdate=1,
makes sure that they are not recreated on subsequent updates or upgrades
after reaching 18.0.
odoo/upgrade#6803

closes odoo/enterprise#75211

closes odoo/enterprise#75249

X-original-commit: 344ff2cc401fc663e05b5b01df1b51abfc8a885a
Related: odoo/odoo#189816
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>
Signed-off-by: Aslam Noman (asno) <asno@odoo.com>
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>
Signed-off-by: Djamel Touati (otd) <otd@odoo.com>